day_01/pyproject.toml
[project]
name = "day-01"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
]
